requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
click==8.1.7
httpx==0.24.1
numpy==1.24.3 